Devices

Braket.Device — Type

Device

Abstract type representing a generic device which tasks (local or managed) may be run on.

Braket.AwsDevice — Type

AwsDevice <: Device

Struct representing an AWS managed device, either simulator or QPU.

Braket.isavailable — Function

isavailable(d::AwsDevice) -> Bool

Is device d currently available to run tasks on.

Braket.search_devices — Function

search_devices(; kwargs...) -> Vector{Dict{String, Any}}

Search all AWS managed devices and filter the results using kwargs.

Valid kwargs are:

arns::Vector{String}: ARNs of devices to search for.

names::Vector{String}: Names of devices to search for.

types::Vector{String}: Types of devices (e.g. QPU or simulator) to search for.

statuses::Vector{String}: Statuses of devices (e.g. "ONLINE" or "OFFLINE") to search for.

provider_names::Vector{String}: Providers of devices to search for.

Braket.get_devices — Function

get_devices(; kwargs...) -> Vector{AwsDevice}

Return all AWS Devices satisfying the filters in kwargs. The devices have their properties populated and a region-appropriate AWSConfig attached.

Valid kwargs are:

arns::Vector{String}: ARNs of devices to search for.

names::Vector{String}: Names of devices to search for.

types::Vector{String}: Types of devices (e.g. QPU or simulator) to search for.

statuses::Vector{String}: Statuses of devices (e.g. "ONLINE" or "OFFLINE") to search for.

provider_names::Vector{String}: Providers of devices to search for.

order_by::String: property used to sort the devices. Default is "name".