Results

Braket.Result — Type

Result

Abstract type representing a measurement to perform on a Circuit.

See also: Expectation, Variance, Sample, Probability, DensityMatrix, and Amplitude.

Braket.Expectation — Type

Expectation <: Result

Struct which represents a expectation measurement on a Circuit.

Braket.Expectation — Method

Expectation(o, targets) -> Expectation
Expectation(o) -> Expectation

Constructs a Expectation of an observable o on qubits targets.

o may be one of:

Any Observable

A String corresponding to an Observable (e.g. "x"`)

A Vector{String} in which each element corresponds to an Observable

targets may be one of:

A QubitSet

A Vector of Ints and/or Qubits

An Int or Qubit

Absent, in which case the observable o will be applied to all qubits provided it is a single qubit observable.

Braket.Expectation — Method

Expectation(c::Circuit, o, targets) -> Circuit
Expectation(c::Circuit, o) -> Circuit

Constructs a Expectation of an observable o on qubits targets and adds it as a result to Circuit c.

o may be one of:

Any Observable

A String corresponding to an Observable (e.g. "x")

A Vector{String} in which each element corresponds to an Observable

targets may be one of:

A QubitSet

A Vector of Ints and/or Qubits

An Int or Qubit

Absent, in which case the observable o will be applied to all qubits provided it is a single qubit observable.

Examples

julia> c = Circuit();

julia> c = H(c, collect(0:10));

julia> c = Expectation(c, Braket.Observables.Z(), 0);

julia> c = Expectation(c, Braket.Observables.X());

Braket.Variance — Type

Variance <: Result

Struct which represents a variance measurement on a Circuit.

Braket.Variance — Method

Variance(o, targets) -> Variance
Variance(o) -> Variance

Constructs a Variance of an observable o on qubits targets.

o may be one of:

Any Observable

A String corresponding to an Observable (e.g. "x"`)

A Vector{String} in which each element corresponds to an Observable

targets may be one of:

A QubitSet

A Vector of Ints and/or Qubits

An Int or Qubit

Absent, in which case the observable o will be applied to all qubits provided it is a single qubit observable.

Braket.Variance — Method

Variance(c::Circuit, o, targets) -> Circuit
Variance(c::Circuit, o) -> Circuit

Constructs a Variance of an observable o on qubits targets and adds it as a result to Circuit c.

o may be one of:

Any Observable

A String corresponding to an Observable (e.g. "x")

A Vector{String} in which each element corresponds to an Observable

targets may be one of:

A QubitSet

A Vector of Ints and/or Qubits

An Int or Qubit

Absent, in which case the observable o will be applied to all qubits provided it is a single qubit observable.

Examples

julia> c = Circuit();

julia> c = H(c, collect(0:10));

julia> c = Variance(c, Braket.Observables.Z(), 0);

julia> c = Variance(c, Braket.Observables.X());

Braket.Sample — Type

Sample <: Result

Struct which represents a sample measurement on a Circuit.

Braket.Sample — Method

Sample(o, targets) -> Sample
Sample(o) -> Sample

Constructs a Sample of an observable o on qubits targets.

o may be one of:

Any Observable

A String corresponding to an Observable (e.g. "x"`)

A Vector{String} in which each element corresponds to an Observable

targets may be one of:

A QubitSet

A Vector of Ints and/or Qubits

An Int or Qubit

Absent, in which case the observable o will be applied to all qubits provided it is a single qubit observable.

Braket.Sample — Method

Sample(c::Circuit, o, targets) -> Circuit
Sample(c::Circuit, o) -> Circuit

Constructs a Sample of an observable o on qubits targets and adds it as a result to Circuit c.

o may be one of:

Any Observable

A String corresponding to an Observable (e.g. "x")

A Vector{String} in which each element corresponds to an Observable

targets may be one of:

A QubitSet

A Vector of Ints and/or Qubits

An Int or Qubit

Absent, in which case the observable o will be applied to all qubits provided it is a single qubit observable.

Examples

julia> c = Circuit();

julia> c = H(c, collect(0:10));

julia> c = Sample(c, Braket.Observables.Z(), 0);

julia> c = Sample(c, Braket.Observables.X());

Braket.Amplitude — Type

Amplitude <: Result

Struct which represents an amplitude measurement on a Circuit.

Braket.Amplitude — Method

Amplitude(states) -> Amplitude

Constructs an Amplitude measurement of states.

states may be one of:

A Vector{String}

A String

All elements of states must be '0' or '1'.

Braket.Amplitude — Method

Amplitude(c::Circuit, states) -> Circuit

Constructs an Amplitude measurement of states and adds it as a result to Circuit c.

states may be of type:

Vector{String}

String

All elements of states must be '0' or '1'.

Examples

julia> c = Circuit();

julia> c = H(c, collect(0:3));

julia> c = Amplitude(c, "0000");

Braket.Probability — Type

Probability <: Result

Struct which represents a probability measurement on a Circuit.

Braket.Probability — Method

Probability(targets) -> Probability
Probability() -> Probability

Constructs a Probability on qubits targets.

targets may be one of:

A QubitSet

A Vector of Ints and/or Qubits

An Int or Qubit

Absent, in which case the measurement will be applied to all qubits.

Braket.Probability — Method

Probability(c::Circuit, targets) -> Circuit
Probability(c::Circuit) -> Circuit

Constructs a Probability measurement on qubits targets and adds it as a result to Circuit c.

targets may be one of:

A QubitSet

A Vector of Ints and/or Qubits

An Int or Qubit

Absent, in which case the measurement will be applied to all qubits of c.

Examples

julia> c = Circuit();

julia> c = H(c, collect(0:3));

julia> c = Probability(c, 2);

Braket.DensityMatrix — Type

DensityMatrix <: Result

Struct which represents a densitymatrix measurement on a Circuit.

Braket.DensityMatrix — Method

DensityMatrix(targets) -> DensityMatrix
DensityMatrix() -> DensityMatrix

Constructs a DensityMatrix on qubits targets.

targets may be one of:

A QubitSet

A Vector of Ints and/or Qubits

An Int or Qubit

Absent, in which case the measurement will be applied to all qubits.

Braket.DensityMatrix — Method

DensityMatrix(c::Circuit, targets) -> Circuit

Constructs a DensityMatrix measurement on qubits targets and adds it as a result to Circuit c.

targets may be one of:

A QubitSet

A Vector of Ints and/or Qubits

An Int or Qubit

Absent, in which case the measurement will be applied to all qubits of c.

Examples

julia> c = Circuit();

julia> c = H(c, collect(0:3));

julia> c = DensityMatrix(c, 2);

Braket.StateVector — Type

StateVector <: Result

Struct which represents a state vector measurement on a Circuit.

Braket.StateVector — Method

StateVector(c::Circuit) -> Circuit

Constructs a StateVector measurement on all qubits of c and adds it as a result to Circuit c.

Examples

julia> c = Circuit();

julia> c = H(c, collect(0:3));

julia> c = StateVector(c);